pyproject.toml
[project]
name = "3d-jigsaw"
version = "0.1.0"
description = "3D jigsaw generator for splitting large STL files into printable parts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.0.0",
    "numpy-stl>=3.2.0",
    "rectpack>=0.2.2",
]
